PartialParser: move some bigint code to separate file; add hitNextAndReset()
The bigint code is unique to jsbn and should be contained in a separate
file so that other libs could possibly be used to mimic the jsbn API if
desired. The code to make this happen would be in separate files that
can be joined at compile time. Eventually, some sort of function could
be implemented to change bigint libs at runtime.
hitNextAndReset() helps to reduce some duplicated code.