remove cacheString() functionality
Was not working properly (BUF2STR lookups always failed) because a
javascript Map that has Uint8Array keys will not check the contents of
the Uint8Array for equality; it will just check object reference
equality. Therefore, this feature needs to be redesigned if desired.

Several possible designs for PartialParser.BUF2STR:
1) A map with hex encoded or base-64 encoded string keys that represent
the Uint8Array key as a string.
2) The Uint8Array object could be hashed and the integer hashcode used
as the map key that maps to a bucket/array of objects with the same hash
code. Then  loop through the objects and compare with the desired
Uint8Array to find a match.