build: re-order closure compiler invocations
The goal of this change is to print helpful closure compiler
warnings/errors earlier/faster in the build process. Minimizing the
src/*.js files first will help to identify the source of the
error/warning when closure compiler prints a message (rather than a line
in a bundled .js file).